+++ eth/abi/bind/template.go
